feat: move new charts into always-on Metrics tab
